+++ lib/column_maker.dart
@@ -1,5 +1,13 @@
 import 'package:sunmi_printer_plus/enums.dart';
 
+///*ColumnMaker*
+/// With this class you can build a column with some text and alignment
+/// This is a good approach if you want to do a recepit like
+/// With [text] , [width] and [align] you can build  a row as you like with any combination of alignment and size<br>
+/// Name  Qtd Value
+///
+/// XXX   2   2.33
+///
 class ColumnMaker {
   String text;
   int width;