+++ lib/enums.dart
@@ -1,6 +1,10 @@
 // ignore_for_file: constant_identifier_names
 library enums;
 
+///*PrinterStatus*
+///
+///This enum will give you the status of the printer.
+///Sometimes the status can be ERROR, but don't worry about this status, always try co print anyway!
 enum PrinterStatus {
   UNKNOWN,
   ERROR,
@@ -15,19 +19,41 @@ enum PrinterStatus {
   NO_BLACK_MARK,
   NO_PRINTER_DETECTED,
   FAILED_TO_UPGRADE_FIRMWARE,
-  EXCEPTION
+  EXCEPTION,
 }
 
-//Enum to set printer mode
-enum PrinterMode { UNKNOWN, NORMAL_MODE, BLACK_LABEL_MODE, LABEL_MODE }
+///*PrinterMode*
+///
+///Enum to set printer mode
+enum PrinterMode {
+  UNKNOWN,
+  NORMAL_MODE,
+  BLACK_LABEL_MODE,
+  LABEL_MODE,
+}
 
-//Enum to set printer aligntment
-enum SunmiPrintAlign { LEFT, CENTER, RIGHT }
+///*SunmiPrintAlign*
+///
+///Enum to set printer aligntment
+enum SunmiPrintAlign {
+  LEFT,
+  CENTER,
+  RIGHT,
+}
 
+///*SunmiQrcodeLevel*
+///
 //Enum to set a QRcode Level (Low to High)
-enum SunmiQrcodeLevel { LEVEL_L, LEVEL_M, LEVEL_Q, LEVEL_H }
+enum SunmiQrcodeLevel {
+  LEVEL_L,
+  LEVEL_M,
+  LEVEL_Q,
+  LEVEL_H,
+}
 
-//Enum to set Barcode Type
+///*SunmiBarcodeType*
+///
+///Enum to set Barcode Type
 enum SunmiBarcodeType {
   UPCA,
   UPCE,
@@ -37,11 +63,25 @@ enum SunmiBarcodeType {
   ITF,
   CODABAR,
   CODE93,
-  CODE128
+  CODE128,
 }
 
-//Enum to set how the thex will be printed in barcode
-enum SunmiBarcodeTextPos { NO_TEXT, TEXT_ABOVE, TEXT_UNDER, BOTH }
+///*SunmiBarcodeTextPos*
+///
+///Enum to set how the thex will be printed in barcode
+enum SunmiBarcodeTextPos {
+  NO_TEXT,
+  TEXT_ABOVE,
+  TEXT_UNDER,
+  BOTH,
+}
 
-//Enum to set font in the printer
-enum SunmiFontSize { XS, SM, MD, LG, XL }
+///*SunmiFontSize*
+///Enum to set font in the printer
+enum SunmiFontSize {
+  XS,
+  SM,
+  MD,
+  LG,
+  XL,
+}