+++ lib/sunmi_printer_plus.dart
@@ -4,9 +4,16 @@ import 'dart:typed_data';
 import 'package:flutter/services.dart';
 import 'package:sunmi_printer_plus/enums.dart';
 import 'package:sunmi_printer_plus/column_maker.dart';
+import 'package:sunmi_printer_plus/sunmi_style.dart';
+export 'package:sunmi_printer_plus/sunmi_style.dart';
 export 'package:sunmi_printer_plus/column_maker.dart';
 export 'package:sunmi_printer_plus/enums.dart';
 
+///*SunmiPrinter*
+///
+///This class is a core of our printer system.
+///With this class you can print everything you like. If you want to print a text, qrcode, barcode, bold text, bigger text or the smallest text possible.
+///Your imagination is the limit!
 class SunmiPrinter {
   static final Map _printerStatus = {
     'ERROR': 'Something went wrong.',
@@ -24,23 +31,45 @@ class SunmiPrinter {
     'EXCEPTION': 'Unknown Error code',
   };
 
+  ///A list to make human read the paper size
+  static final List<int> _paperize = [80, 58];
+
+  ///*sunmi_printer_plus
+  ///
+  //A callable method to start the comunitation with the native code!
   static const MethodChannel _channel = MethodChannel('sunmi_printer_plus');
 
+  ///*bindingPrinter*
+  ///
+  ///This method will intializate the printer to start the whole print.
+  ///This method *Must* be executed before any other print command.
   static Future<bool?> bindingPrinter() async {
     final bool? status = await _channel.invokeMethod('BIND_PRINTER_SERVICE');
     return status;
   }
+
+  ///*unbindingPrinter*
+  ///
+  ///This method is the opposite of [bindingPrinter].
+  ///This will unbind or 'close' the connection with the printer, and must be the last execution.
 
   static Future<bool?> unbindingPrinter() async {
     final bool? status = await _channel.invokeMethod('UNBIND_PRINTER_SERVICE');
     return status;
   }
 
+  ///*initPrinter*
+  ///
+  ///This method will reset the printer styles, font and everything that can change your text, but the method will *NOT* erase the buffer inside.
   static Future<bool?> initPrinter() async {
     final bool? status = await _channel.invokeMethod('INIT_PRINTER');
     return status;
   }
 
+  ///*getPrinterStatus*
+  ///
+  ///This method will give you the status of the printer.
+  ///Sometimes the printer can give you an error, so, try to print anyway.
   static Future<PrinterStatus> getPrinterStatus() async {
     final String? status = await _channel.invokeMethod('GET_UPDATE_PRINTER');
     switch (status) {
@@ -75,12 +104,19 @@ class SunmiPrinter {
     }
   }
 
+  ///*getPrinterStatusWithVerbose*
+  ///
+  ///Almos the same of  [getPrinterStatus] , but will return a text explaned
+  ///@see the _printerStatus map!
   static Future<String?> getPrinterStatusWithVerbose() async {
     final String? status = await _channel.invokeMethod('GET_UPDATE_PRINTER');
     final statusMsg = _printerStatus[status];
     return statusMsg;
   }
 
+  ///*getPrinterMode*
+  ///
+  ///This method will return what mode your printer to print in one way or other, like label mode or normal mode
   static Future<PrinterMode> getPrinterMode() async {
     final String mode = await _channel.invokeMethod('GET_PRINTER_MODE');
     switch (mode) {
@@ -95,11 +131,34 @@ class SunmiPrinter {
     }
   }
 
-  static Future<void> printText(String text) async {
+  ///*printText*
+  ///
+  ///This method will print a simple text in your printer
+  /// With the [SunmiStyle] you can put in one line, the size, alignment and bold
+  static Future<void> printText(String text, {SunmiStyle? style}) async {
+    if (style != null) {
+      if (style.align != null) {
+        await setAlignment(style.align!);
+      }
+
+      if (style.fontSize != null) {
+        await setFontSize(style.fontSize!);
+      }
+
+      if (style.bold != null) {
+        if (style.bold == true) {
+          await bold();
+        }
+      }
+    }
     Map<String, dynamic> arguments = <String, dynamic>{"text": '$text\n'};
     await _channel.invokeMethod("PRINT_TEXT", arguments);
+    await initPrinter();
   }
 
+  ///*printRow*
+  ///
+  ///This method will print a row based in a list of [ColumnMaker].
   static Future<void> printRow({required List<ColumnMaker> cols}) async {
     final _jsonCols = List<Map<String, String>>.from(
         cols.map<Map<String, String>>((ColumnMaker col) => col.toJson()));
@@ -109,11 +168,18 @@ class SunmiPrinter {
     await _channel.invokeMethod("PRINT_ROW", arguments);
   }
 
+  ///*printRawData*
+  ///
+  ///With this method you can print a raw data, or a data that was made with some esc/pos package to simplify your calls
+  ///*This is good if you have another print method that give you a [List<int>] that you can convert to esc/pos here
   static Future<void> printRawData(Uint8List data) async {
     Map<String, dynamic> arguments = <String, dynamic>{"data": data};
     await _channel.invokeMethod("RAW_DATA", arguments);
   }
 
+  ///*printQRCode*
+  ///
+  ///With this method you can print a qrcode with some errorLevel and size.
   static Future<void> printQRCode(String data,
       {int size = 5,
       SunmiQrcodeLevel errorLevel = SunmiQrcodeLevel.LEVEL_H}) async {
@@ -141,6 +207,9 @@ class SunmiPrinter {
     await _channel.invokeMethod("PRINT_QRCODE", arguments);
   }
 
+  ///*printBarCode*
+  ///
+  ///With this method you can print a barcode with any type discribed below or in the enum section
   static Future<void> printBarCode(String data,
       {SunmiBarcodeType barcodeType = SunmiBarcodeType.CODE128,
       int height = 162,
@@ -203,11 +272,17 @@ class SunmiPrinter {
     await _channel.invokeMethod("PRINT_BARCODE", arguments);
   }
 
+  ///*lineWrap*
+  ///
+  ///With this method you can jump N lines in your pinter to make some spaces between sections
   static Future<void> lineWrap(int lines) async {
     Map<String, dynamic> arguments = <String, dynamic>{"lines": lines};
     await _channel.invokeMethod("LINE_WRAP", arguments);
   }
 
+  ///*line*
+  ///
+  ///With this method you can draw a line to divide sections.
   static Future<void> line({
     String ch = '-',
     int len = 31,
@@ -216,17 +291,28 @@ class SunmiPrinter {
     await printText(List.filled(len, ch[0]).join());
   }
 
+  ///*bold*
+  ///
+  ///With this method you can bold a string very easy, just put this method before a [printText] and everyting after this mehod will be bold
+
   static Future<void> bold() async {
     final List<int> boldOn = [27, 69, 1];
 
     await printRawData(Uint8List.fromList(boldOn));
   }
 
+  ///*resetBold*
+  ///
+  ///This method will just reset the bold to a normal font weight
   static Future<void> resetBold() async {
     final List<int> boldOff = [27, 69, 0];
 
     await printRawData(Uint8List.fromList(boldOff));
   }
+
+  ///*setAlignment*
+  ///
+  ///With this method you can align your text in three ways, like LEFT, RIGHT and CENTER.
 
   static Future<void> setAlignment(SunmiPrintAlign alignment) async {
     late int value;
@@ -247,25 +333,51 @@ class SunmiPrinter {
     await _channel.invokeMethod("SET_ALIGNMENT", arguments);
   }
 
+  ///*printImage*
+  ///
+  ///With this method you can print an image in your printer.
+  ///Just follow the examples that you can print even an image from web or an asset inside your project
   static Future<void> printImage(Uint8List img) async {
     Map<String, dynamic> arguments = <String, dynamic>{};
     arguments.putIfAbsent("bitmap", () => img);
     await _channel.invokeMethod("PRINT_IMAGE", arguments);
   }
 
+  ///*startTransactionPrint*
+  ///
+  ///If you want to print in one transaction, you can start the transaction, build your print commands without send to the buffer
   static Future<void> startTransactionPrint([bool clear = false]) async {
     Map<String, dynamic> arguments = <String, dynamic>{"clearEnter": clear};
     await _channel.invokeMethod("ENTER_PRINTER_BUFFER", arguments);
   }
 
+  ///*submitTransactionPrint*
+  ///
+  ///This method will submit your transaction to the bufffer
   static Future<void> submitTransactionPrint() async {
     await _channel.invokeMethod("COMMIT_PRINTER_BUFFER");
   }
+
+  ///*cut*
+  ///
+  ///This method will  cut the paper
+  static Future<void> cut() async {
+    await _channel.invokeMethod("CUT_PAPER");
+  }
+
+  ///*exitTransactionPrint*
+  ///
+  ///This method will close the transaction
 
   static Future<void> exitTransactionPrint([bool clear = true]) async {
     Map<String, dynamic> arguments = <String, dynamic>{"clearExit": clear};
     await _channel.invokeMethod("EXIT_PRINTER_BUFFER", arguments);
   }
+
+  ///*setFontSize*
+  ///
+  ///This method will change the fontsize , between extra small and extra large.
+  ///You can see the sizes below or in the enum file.
 
   static Future<void> setFontSize(SunmiFontSize _size) async {
     int _fontSize = 24;
@@ -291,16 +403,33 @@ class SunmiPrinter {
     await _channel.invokeMethod("FONT_SIZE", arguments);
   }
 
+  ///*resetFontSize*
+  ///
+  ///This method will reset the font size to the medium (default) size
   static Future<void> resetFontSize() async {
     Map<String, dynamic> arguments = <String, dynamic>{"size": 24};
     await _channel.invokeMethod("FONT_SIZE", arguments);
   }
 
-  static Future<void> startLabelPrint() async {
-    await _channel.invokeMethod("LABEL_LOCATE");
+  ///*paperSize*
+  ///
+  /// Get the paper size , because they can change the printers between 56 and 80mm
+  static Future<int> paperSize() async {
+    int? _size = await _channel.invokeMethod("PAPER_SIZE");
+    return _paperize[_size ?? 0];
   }
 
-  static Future<void> exitLabelPrint() async {
-    await _channel.invokeMethod("LABEL_OUTPUT");
+  ///*serialNumber*
+  ///
+  /// Get the serial number
+  static Future<String> serialNumber() async {
+    return await _channel.invokeMethod("PRINTER_SERIAL_NUMBER");
+  }
+
+  ///*printerVersion*
+  ///
+  /// Get the printer's version
+  static Future<String> printerVersion() async {
+    return await _channel.invokeMethod("PRINTER_VERSION");
   }
 }