+++ lib/sunmi_style.dart
@@ -1,0 +1,14 @@
+import 'package:sunmi_printer_plus/enums.dart';
+
+///*SunmiStyle*
+///
+///With this class you can build your own layout to print some text.
+///The thext can come already with [align], [fontSize] and *bold*, and then you don't need to type 3 commands to do the same THING!
+
+class SunmiStyle {
+  SunmiFontSize? fontSize;
+  SunmiPrintAlign? align;
+  bool? bold;
+
+  SunmiStyle({this.fontSize, this.align, this.bold});
+}